rename labels ➡️ resources
